auth_lib/pyproject.toml
[project]
name = "a2a-auth-lib"
version = "0.1.0"
dependencies = [
    "requests",
    "PyJWT",
]
